@@ -10,7 +10,7 @@
 <h2 class="text-info">Item Details</h2>
 <br />
 
-<form method="post" asp-action="Create">
+<form method="post">
     <div class="border backgroundWhite row">
 
         <div asp-validation-summary="ModelOnly" class="text-danger"></div>
